refactor: move files to directories
