+++ src/index.ts
@@ -1,13 +1,13 @@
-export * from './array.js';
-export * from './env.js';
-export * from './error.js';
-export * from './function.js';
-export * from './guards.js';
-export * from './nullable.js';
-export * from './number.js';
-export * from './object.js';
-export * from './package.js';
-export * from './regex.js';
-export * from './string.js';
-export * from './types.js';
-export * from './values.js';
+export * from './array/array.js';
+export * from './env/env.js';
+export * from './error/error.js';
+export * from './function/function.js';
+export * from './guards/guards.js';
+export * from './nullable/nullable.js';
+export * from './number/number.js';
+export * from './object/object.js';
+export * from './package/package.js';
+export * from './regex/regex.js';
+export * from './string/string.js';
+export * from './types/types.js';
+export * from './values/values.js';
