+++ src/internals/string.ts
@@ -1,3 +1,3 @@
-import { testRegex } from '../regex.js';
+import { testRegex } from '../regex/regex.js';
 
 export const isWordOrWords = (x: string) => testRegex(/(^.+ .)|(^\S+$)/, x);
