+++ src/object/object.spec.ts
