+++ src/object/object.ts
