+++ src/string/string.spec.ts
