+++ src/string/string.ts
@@ -1,7 +1,7 @@
 import { isNull, isString } from './guards.js';
-import { isWordOrWords } from './internals/string.js';
-import { hasInvalidCasing, tokenize } from './internals/tokenize.js';
-import type { Nullish, Sensitivity } from './types.js';
+import { isWordOrWords } from '../internals/string.js';
+import { hasInvalidCasing, tokenize } from '../internals/tokenize.js';
+import type { Nullish, Sensitivity } from '../types/types.js';
 
 export const toSpaceCase = tokenize;
